src/pyAdvancedLic_Server/app/requirements.txt
sqlalchemy
asyncpg
fastapi
python-dotenv
pydantic
uvicorn
redis
fastapi-utils
aiologger
aiofiles